Rename conversion to xarray, Rename Nodes to Transpose and ToDataFrame, Clean-up and introduce ready method to check ports.
